move ImageGallery component into its own directory
